+++ packages/gen-typescript-declarations/CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+* Added `--googModules` flag, for generating declarations compatible
+  with `goog.module`-style modules.
+* Added `--hideWarnings` flag, for hiding warning-level severity diagnostics.
+* Generate declarations with tslint disables for relevant rules, like variable
+  names, and no-any.
+
 <!-- Add new, unreleased changes here. -->
 
 ## [1.5.1] - 2018-08-25